# openapi/api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mock Cloud API",
    "description": "\n    ## Mock Cloud API\n    \n    A comprehensive mock cloud backend API for testing, development, and SDK generation.\n    \n    ### Features\n    * **Virtual Machines** - Create, manage, and monitor VMs\n    * **Volumes** - Block storage management with VM attachment\n    * **Security Groups** - Network security and access control\n    * **Environments** - Network isolation with CIDR management\n    * **Async Operations** - Background task processing with Celery\n    \n    ### Use Cases\n    * **Development & Testing** - Mock cloud infrastructure\n    * **SDK Development** - Generate client libraries\n    * **Terraform Providers** - Infrastructure as code testing\n    * **CI/CD Pipelines** - Automated testing environments\n    \n    ### Authentication\n    No authentication required - suitable for development and testing.\n    \n    ### Rate Limiting\n    No rate limiting applied - suitable for load testing.\n    \n    ### SDK Generation\n    This API is designed to work seamlessly with OpenAPI code generators:\n    * Python SDK generation\n    * Go SDK generation  \n    * Node.js SDK generation\n    * Terraform provider development\n    ",
    "contact": {
      "name": "Mock Cloud API Support",
      "email": "support@mockcloud.local"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
...
  "paths": {
...
    "/api/v1/volumes/": {
      "post": {
        "tags": [
          "volumes",
          "volumes"
        ],
        "summary": "Create Volume",
        "description": "Create a new volume.\n\nCreates a block storage volume that can be attached to VMs.\nVolume creation is asynchronous and takes 30-60 seconds.\n\n**Parameters:**\n* `name`: Unique name for the volume\n* `size_gb`: Size in gigabytes\n* `environment_id`: ID of the environment to create the volume in\n\n**Returns:**\nTask response with volume ID and task tracking information",
        "operationId": "create_volume_api_v1_volumes__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VolumeCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "volumes",
          "volumes"
        ],
        "summary": "Get All Volumes",
        "description": "Get all volumes with pagination.\n\nRetrieves a paginated list of all volumes.\nUseful for listing available volumes and implementing\npagination in client applications.\n\n**Parameters:**\n* `skip`: Number of volumes to skip (for pagination)\n* `limit`: Maximum number of volumes to return\n\n**Returns:**\nPaginated list of volumes with total count",
        "operationId": "get_all_volumes_api_v1_volumes__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/volumes/{volume_id}": {
      "get": {
        "tags": [
          "volumes",
          "volumes"
        ],
        "summary": "Get Volume",
        "description": "Get a specific volume by ID.\n\nRetrieves detailed information about a volume including\nits size, status, and attachment information.\n\n**Parameters:**\n* `volume_id`: Unique identifier for the volume\n\n**Returns:**\nComplete volume information with metadata",
        "operationId": "get_volume_api_v1_volumes__volume_id__get",
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Volume Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "volumes",
          "volumes"
        ],
        "summary": "Delete Volume",
        "description": "Delete a volume by ID.\n\nStarts the deletion process for a volume.\nVolume deletion is asynchronous and may take time\nif the volume is currently attached to a VM.\n\n**Parameters:**\n* `volume_id`: Unique identifier for the volume to delete\n\n**Returns:**\nTask response with deletion task information\n\n**Note:**\nThis operation will fail if the volume is currently attached to a VM.",
        "operationId": "delete_volume_api_v1_volumes__volume_id__delete",
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Volume Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/volumes/{volume_id}/attach/{vm_id}": {
      "post": {
        "tags": [
          "volumes",
          "volumes"
        ],
        "summary": "Attach Volume To Vm",
        "description": "Attach a volume to a VM.\n\nAttaches a volume to a specific VM, making it available\nfor use by the VM. The volume status will change to 'in_use'.\n\n**Parameters:**\n* `volume_id`: Unique identifier for the volume\n* `vm_id`: Unique identifier for the VM to attach to\n\n**Returns:**\nSuccess message confirming attachment\n\n**Note:**\nA volume can only be attached to one VM at a time.",
        "operationId": "attach_volume_to_vm_api_v1_volumes__volume_id__attach__vm_id__post",
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Volume Id"
            }
          },
          {
            "name": "vm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/volumes/{volume_id}/detach": {
      "post": {
        "tags": [
          "volumes",
          "volumes"
        ],
        "summary": "Detach Volume From Vm",
        "description": "Detach a volume from its current VM.\n\nDetaches a volume from the VM it's currently attached to.\nThe volume status will change to 'available' and can be\nattached to another VM.\n\n**Parameters:**\n* `volume_id`: Unique identifier for the volume to detach\n\n**Returns:**\nSuccess message confirming detachment",
        "operationId": "detach_volume_from_vm_api_v1_volumes__volume_id__detach_post",
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Volume Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ResourceStatus": {
        "type": "string",
        "enum": [
          "pending",
          "creating",
          "active",
          "failed",
          "deleting",
          "deleted"
        ],
        "title": "ResourceStatus"
      },
...
      "TaskResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "resource_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Type"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status",
          "message"
        ],
        "title": "TaskResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VolumeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "size_gb": {
            "type": "integer",
            "maximum": 10000.0,
            "exclusiveMinimum": 0.0,
            "title": "Size Gb"
          },
          "environment_id": {
            "type": "integer",
            "title": "Environment Id"
          },
          "vm_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vm Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "size_gb",
          "environment_id"
        ],
        "title": "VolumeCreate"
      },
      "VolumeListResponse": {
        "properties": {
          "volumes": {
            "items": {
              "$ref": "#/components/schemas/VolumeResponse"
            },
            "type": "array",
            "title": "Volumes"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "volumes",
          "total"
        ],
        "title": "VolumeListResponse"
      },
      "VolumeResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "size_gb": {
            "type": "integer",
            "maximum": 10000.0,
            "exclusiveMinimum": 0.0,
            "title": "Size Gb"
          },
          "environment_id": {
            "type": "integer",
            "title": "Environment Id"
          },
          "vm_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vm Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/VolumeStatus"
          },
          "resource_status": {
            "$ref": "#/components/schemas/ResourceStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "size_gb",
          "environment_id",
          "id",
          "status",
          "resource_status",
          "created_at"
        ],
        "title": "VolumeResponse"
      },
      "VolumeStatus": {
        "type": "string",
        "enum": [
          "pending",
          "creating",
          "available",
          "in_use",
          "failed",
          "deleting",
          "deleted"
        ],
        "title": "VolumeStatus"
      }
    }
  }
}